+++ src/answers/__tests__/answers.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "@jest/globals";
-import { DatePickerAnswerSchema, DateRangeAnswerSchema } from '../dateAnswers';
+import { DateAnswerSchema, DateRangeAnswerSchema } from '../dateAnswers';
 import { FilteredSearchAnswerSchema, TypeaheadSearchAnswerSchema } from '../graphQLAnswers';
 import { CheckboxesAnswerSchema, RadioButtonsAnswerSchema, SelectBoxAnswerSchema } from '../optionBasedAnswers';
 import {
@@ -7,6 +7,7 @@ import {
   CurrencyAnswerSchema,
   EmailAnswerSchema,
   NumberAnswerSchema,
+  NumberRangeAnswerSchema,
   TextAnswerSchema,
   TextAreaAnswerSchema,
   URLAnswerSchema
@@ -40,12 +41,12 @@ describe('Answer Type Validations', () => {
     expect(() => CurrencyAnswerSchema.parse(invalidData)).toThrow();
   });
 
-  it('should validate DatePickerAnswer', () => {
-    const validData = { type: 'datePicker', answer: '2023-10-01', meta: { schemaVersion: CURRENT_SCHEMA_VERSION } };
-    expect(() => DatePickerAnswerSchema.parse(validData)).not.toThrow();
+  it('should validate DateAnswer', () => {
+    const validData = { type: 'date', answer: '2023-10-01', meta: { schemaVersion: CURRENT_SCHEMA_VERSION } };
+    expect(() => DateAnswerSchema.parse(validData)).not.toThrow();
 
-    const invalidData = { type: 'datePicker', answer: 12345, meta: { schemaVersion: CURRENT_SCHEMA_VERSION } };
-    expect(() => DatePickerAnswerSchema.parse(invalidData)).toThrow();
+    const invalidData = { type: 'date', answer: 12345, meta: { schemaVersion: CURRENT_SCHEMA_VERSION } };
+    expect(() => DateAnswerSchema.parse(invalidData)).toThrow();
   });
 
   it('should validate DateRangeAnswer', () => {
@@ -86,6 +87,22 @@ describe('Answer Type Validations', () => {
 
     const invalidData = { type: 'number', answer: '42', meta: { schemaVersion: CURRENT_SCHEMA_VERSION } };
     expect(() => NumberAnswerSchema.parse(invalidData)).toThrow();
+  });
+
+  it('should validate NumberRangeAnswer', () => {
+    const validData = {
+      type: 'numberRange',
+      answer: { start: 1, end: 10 },
+      meta: { schemaVersion: CURRENT_SCHEMA_VERSION }
+    };
+    expect(() => NumberRangeAnswerSchema.parse(validData)).not.toThrow();
+
+    const invalidData = {
+      type: 'numberRange',
+      answer: { start: '1', end: 10 },
+      meta: { schemaVersion: CURRENT_SCHEMA_VERSION }
+    };
+    expect(() => NumberRangeAnswerSchema.parse(invalidData)).toThrow();
   });
 
   it('should validate RadioButtonsAnswer', () => {
@@ -140,8 +157,46 @@ describe('Answer Type Validations', () => {
     const validData = {
       type: 'table',
       answer: [
-        { type: 'text', answer: 'Row 1', meta: { schemaVersion: CURRENT_SCHEMA_VERSION } },
-        { type: 'number', answer: 42, meta: { schemaVersion: CURRENT_SCHEMA_VERSION } },
+        {
+          columns: [
+            {
+              heading: "Name",
+              content: {
+                type: 'text',
+                answer: 'Leia Organa',
+                meta: { schemaVersion: CURRENT_SCHEMA_VERSION }
+              }
+            },
+            {
+              heading: "Age",
+              content: {
+                type: 'number',
+                answer: 28,
+                meta: { schemaVersion: CURRENT_SCHEMA_VERSION }
+              }
+            }
+          ]
+        },
+        {
+          columns: [
+            {
+              heading: "Name",
+              content: {
+                type: 'text',
+                answer: 'Han Solo',
+                meta: { schemaVersion: CURRENT_SCHEMA_VERSION }
+              }
+            },
+            {
+              heading: "Age",
+              content: {
+                type: 'number',
+                answer: 35,
+                meta: { schemaVersion: CURRENT_SCHEMA_VERSION }
+              }
+            }
+          ]
+        }
       ],
       meta: { schemaVersion: CURRENT_SCHEMA_VERSION }
     };