+++ src/questions/index.ts
@@ -4,11 +4,12 @@ import {
   CurrencyQuestionSchema,
   EmailQuestionSchema,
   NumberQuestionSchema,
+  NumberRangeQuestionSchema,
   TextAreaQuestionSchema,
   TextQuestionSchema,
   URLQuestionSchema,
 } from "./primitiveQuestions";
-import { DatePickerQuestionSchema, DateRangeQuestionSchema } from "./dateQuestions";
+import { DateQuestionSchema, DateRangeQuestionSchema } from "./dateQuestions";
 import { CheckboxesQuestionSchema, RadioButtonsQuestionSchema, SelectBoxQuestionSchema } from './optionBasedQuestions';
 import { FilteredSearchQuestionSchema, TypeaheadSearchQuestionSchema } from './graphQLQuestions';
 import { TableQuestionSchema } from './tableQuestions';
@@ -27,7 +28,7 @@ export const AnyQuestionSchema = z.discriminatedUnion('type', [
   BooleanQuestionSchema,
   CheckboxesQuestionSchema,
   CurrencyQuestionSchema,
-  DatePickerQuestionSchema,
+  DateQuestionSchema,
   DateRangeQuestionSchema,
   EmailQuestionSchema,
   FilteredSearchQuestionSchema,
@@ -46,11 +47,12 @@ export const QuestionSchemaMap: Record<z.infer<typeof QuestionTypesEnum>, z.ZodT
   boolean: BooleanQuestionSchema,
   checkBoxes: CheckboxesQuestionSchema,
   currency: CurrencyQuestionSchema,
-  datePicker: DatePickerQuestionSchema,
+  date: DateQuestionSchema,
   dateRange: DateRangeQuestionSchema,
   email: EmailQuestionSchema,
   filteredSearch: FilteredSearchQuestionSchema,
   number: NumberQuestionSchema,
+  numberRange: NumberRangeQuestionSchema,
   radioButtons: RadioButtonsQuestionSchema,
   selectBox: SelectBoxQuestionSchema,
   table: TableQuestionSchema,