+++ src/questions/tableQuestions.ts
@@ -8,7 +8,7 @@ import {
   TextQuestionSchema,
   URLQuestionSchema,
 } from "./primitiveQuestions";
-import { DatePickerQuestionSchema, DateRangeQuestionSchema } from "./dateQuestions";
+import { DateQuestionSchema, DateRangeQuestionSchema } from "./dateQuestions";
 import { CheckboxesQuestionSchema, RadioButtonsQuestionSchema, SelectBoxQuestionSchema } from './optionBasedQuestions';
 import { FilteredSearchQuestionSchema, TypeaheadSearchQuestionSchema } from './graphQLQuestions';
 import { QuestionSchema } from "./question";
@@ -18,7 +18,7 @@ export const AnyTableColumnQuestionSchema = z.discriminatedUnion('type', [
   BooleanQuestionSchema,
   CheckboxesQuestionSchema,
   CurrencyQuestionSchema,
-  DatePickerQuestionSchema,
+  DateQuestionSchema,
   DateRangeQuestionSchema,
   EmailQuestionSchema,
   FilteredSearchQuestionSchema,
@@ -31,17 +31,22 @@ export const AnyTableColumnQuestionSchema = z.discriminatedUnion('type', [
   URLQuestionSchema
 ]);
 
+export const TableColumn = z.object({
+  heading: z.string().optional(),                           // The heading of the column
+  content: AnyTableColumnQuestionSchema,                    // The question for the column
+});
+
 // Table question and answer
 export const TableQuestionSchema = QuestionSchema.merge(z.object({
   type: z.literal('table'),                                 // The type of question
-  columns: z.array(AnyTableColumnQuestionSchema),           // The columns of the table (note: tables cannot be nested)
+  columns: z.array(TableColumn),                            // The columns of the table
   attributes: z.object({
     canAddRows: z.boolean().optional(),                     // Whether to allow adding rows (default is true)
     canRemoveRows: z.boolean().optional(),                  // Whether to allow removing rows (default is true)
     initialRows: z.number().optional(),                     // The initial number of rows (default is 1)
     maxRows: z.number().optional(),                         // The maximum number of rows (no default)
     minRows: z.number().optional()                          // The minimum number of rows (no default)
-  })
+  }).optional(),
 }));
 
 // This will ensure that object validations are against the Zod schemas defined above